organized_tutorial/rag-chatbot-test/requirements.txt
requests==2.31.0
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.0.1
numpy==1.24.3
openai==1.3.0
langchain==0.0.350
chromadb==0.4.18
sentence-transformers==2.2.2
